Camera CR fixes

+++ libs/drivers/camera/command_factory.cpp
@@ -6,8 +6,8 @@ using namespace devices::camera;
 void CommandFactory::BuildAck(Frame cmd, CameraCmd cmdAck, uint8_t packageIdLow, uint8_t packageIdHigh)
 {
     cmd[0] = CommandPrefix;
-    cmd[1] = (uint8_t)CameraCmd::Ack;
-    cmd[2] = (uint8_t)cmdAck;
+    cmd[1] = static_cast<uint8_t>(CameraCmd::Ack);
+    cmd[2] = static_cast<uint8_t>(cmdAck);
     cmd[3] = 0x00;
     cmd[4] = packageIdLow;
     cmd[5] = packageIdHigh;
@@ -16,7 +16,7 @@ void CommandFactory::BuildAck(Frame cmd, CameraCmd cmdAck, uint8_t packageIdLow,
 void CommandFactory::BuildSync(Frame cmd)
 {
     cmd[0] = CommandPrefix;
-    cmd[1] = (uint8_t)CameraCmd::Sync;
+    cmd[1] = static_cast<uint8_t>(CameraCmd::Sync);
     cmd[2] = 0x00;
     cmd[3] = 0x00;
     cmd[4] = 0x00;
@@ -26,38 +26,38 @@ void CommandFactory::BuildSync(Frame cmd)
 void CommandFactory::BuildReset(Frame cmd, CameraResetType resetType)
 {
     cmd[0] = CommandPrefix;
-    cmd[1] = (uint8_t)CameraCmd::Reset;
-    cmd[2] = (uint8_t)resetType;
+    cmd[1] = static_cast<uint8_t>(CameraCmd::Reset);
+    cmd[2] = static_cast<uint8_t>(resetType);
     cmd[3] = 0x00;
     cmd[4] = 0x00;
-    cmd[5] = (uint8_t)0xFF;
+    cmd[5] = 0xFF;
 }
 
 void CommandFactory::BuildInitRAW(Frame cmd, CameraRAWImageFormat format, CameraRAWResolution rawResolution)
 {
     cmd[0] = CommandPrefix;
-    cmd[1] = (uint8_t)CameraCmd::Initial;
+    cmd[1] = static_cast<uint8_t>(CameraCmd::Initial);
     cmd[2] = 0x00;
-    cmd[3] = (uint8_t)format;
-    cmd[4] = (uint8_t)rawResolution;
+    cmd[3] = static_cast<uint8_t>(format);
+    cmd[4] = static_cast<uint8_t>(rawResolution);
     cmd[5] = 0x00;
 }
 
 void CommandFactory::BuildInitJPEG(Frame cmd, CameraJPEGResolution jpegResolution)
 {
     cmd[0] = CommandPrefix;
-    cmd[1] = (uint8_t)CameraCmd::Initial;
+    cmd[1] = static_cast<uint8_t>(CameraCmd::Initial);
     cmd[2] = 0x00;
     cmd[3] = 0x07;
     cmd[4] = 0x07;
-    cmd[5] = (uint8_t)jpegResolution;
+    cmd[5] = static_cast<uint8_t>(jpegResolution);
 }
 
 void CommandFactory::BuildGetPicture(Frame cmd, CameraPictureType::Enum type)
 {
     cmd[0] = CommandPrefix;
-    cmd[1] = (uint8_t)CameraCmd::GetPicture;
-    cmd[2] = (uint8_t)type;
+    cmd[1] = static_cast<uint8_t>(CameraCmd::GetPicture);
+    cmd[2] = static_cast<uint8_t>(type);
     cmd[3] = 0x00;
     cmd[4] = 0x00;
     cmd[5] = 0x00;
@@ -66,8 +66,8 @@ void CommandFactory::BuildGetPicture(Frame cmd, CameraPictureType::Enum type)
 void CommandFactory::BuildSnapshot(Frame cmd, CameraSnapshotType type)
 {
     cmd[0] = CommandPrefix;
-    cmd[1] = (uint8_t)CameraCmd::Snapshot;
-    cmd[2] = (uint8_t)type;
+    cmd[1] = static_cast<uint8_t>(CameraCmd::Snapshot);
+    cmd[2] = static_cast<uint8_t>(type);
     cmd[3] = 0x00;
     cmd[4] = 0x00;
     cmd[5] = 0x00;
@@ -82,10 +82,10 @@ bool CommandFactory::BuildSetPackageSize(Frame cmd, uint16_t packageSize)
     }
 
     cmd[0] = CommandPrefix;
-    cmd[1] = (uint8_t)CameraCmd::SetPackageSize;
+    cmd[1] = static_cast<uint8_t>(CameraCmd::SetPackageSize);
     cmd[2] = 0x08;
-    cmd[3] = (uint8_t)(packageSize & 0xFF);
-    cmd[4] = (uint8_t)((packageSize >> 8) & 0xFF);
+    cmd[3] = static_cast<uint8_t>((packageSize & 0xFF));
+    cmd[4] = static_cast<uint8_t>(((packageSize >> 8) & 0xFF));
     cmd[5] = 0x00;
 
     return true;
@@ -94,7 +94,7 @@ bool CommandFactory::BuildSetPackageSize(Frame cmd, uint16_t packageSize)
 void CommandFactory::BuildSetBaudRate(Frame cmd, uint8_t firstDivider, uint8_t secondDivider)
 {
     cmd[0] = CommandPrefix;
-    cmd[1] = (uint8_t)CameraCmd::SetBaudRate;
+    cmd[1] = static_cast<uint8_t>(CameraCmd::SetBaudRate);
     cmd[2] = firstDivider;
     cmd[3] = secondDivider;
     cmd[4] = 0x00;
